Split out utils and dag_utils
