+++ src/quartic/pipeline/validator/utils.py
@@ -1,37 +1,7 @@
-import importlib.util
-import pprint
 import json
 from collections import defaultdict
-import os.path
-import networkx as nx
+import pprint
 from networkx.drawing.nx_agraph import write_dot
-from quartic.utils import QuarticException
-
-from ..step import Step
-from ..dataset import Dataset
-
-def build_dag(steps, default_namespace):
-    assert steps
-    datasets = set()
-    for step in steps:
-        for ds in step.datasets():
-            datasets.add(ds.fully_qualified(default_namespace))
-
-    g = nx.DiGraph()
-    for ds in datasets:
-        g.add_node(ds)
-
-    for step in steps:
-        for i in step.inputs():
-            for o in step.outputs():
-                g.add_edge(i.fully_qualified(default_namespace),
-                           o.fully_qualified(default_namespace), label=step.name, step=step)
-    return g
-
-def check_dag(dag):
-    if not nx.is_directed_acyclic_graph(dag):
-        raise QuarticException("graph is not a dag")
-
 # https://stackoverflow.com/questions/2892931/longest-common-substring-from-more-than-two-strings-python
 def common_prefix(strings):
     """ Find the longest string that is a prefix of all the strings.
@@ -117,83 +87,3 @@ def save_json(dag, fname):
         ]
     }
     json.dump(blob, open(fname, "w"), indent=1)
-
-def load_resume_file(resume_file, default_namespace):
-    if resume_file and os.path.exists(resume_file):
-        datasets = [Dataset.parse(s).fully_qualified(default_namespace)
-                    for s in json.load(open(resume_file))]
-
-
-        datasets = set(datasets)
-        print("skipping datasets: {}".format(pprint.pformat(datasets)))
-        return datasets
-    else:
-        print("Resume file not found. Assuming empty.")
-        return set()
-
-def get_module_specs():
-    module_specs = []
-    for f in os.listdir("."):
-        if f.endswith(".py"):
-            module_specs.append(
-                importlib.util.spec_from_file_location(f.strip('.py'), os.path.abspath(f))
-                )
-    assert module_specs
-    return module_specs
-
-def validate():
-    steps = []
-    modules = get_module_specs()
-    for mspec in modules:
-        m = importlib.util.module_from_spec(mspec)
-        mspec.loader.exec_module(m)
-        for k, v in m.__dict__.items():
-            if isinstance(v, Step):
-                steps.append(v)
-
-    # build the DAG and check it
-    dag = build_dag(steps, "local-testing")
-    return check_dag(dag)
-
-def guff_validate(action='graphviz'):
-    steps = []
-    modules = get_module_specs()
-    for module in modules:
-        m = importlib.import_module(module)
-        for k, v in m.__dict__.items():
-            if isinstance(v, Step):
-                steps.append(v)
-
-    # build the DAG and check it
-    dag = build_dag(steps, "local-testing")
-    check_dag(dag)
-
-    # build list of raw datasets and ones to materialise (tsorted)
-    raw_datasets = []
-    materialise_datasets = []
-    for node in nx.topological_sort(dag):
-        if dag.in_degree(node) == 0:
-            raw_datasets.append(node)
-        else:
-            materialise_datasets.append(node)
-
-    # build list of steps to run in topological order
-    run_steps = []
-    for ds in materialise_datasets:
-        edges = dag.in_edges(ds, True)
-        steps = set()
-        for edge in edges:
-            steps.add(edge[2]["step"])
-
-        assert len(steps) == 1
-        run_steps.append(steps.pop())
-
-    if action == "graphviz":
-        save_graphviz(dag, "graph.dot")
-    elif action == "json":
-        save_json(dag, "graph.json")
-    elif action == "explain":
-        pprint.pprint({
-            "raw": raw_datasets,
-            "materialize": materialise_datasets
-        })
